Fixed "Shell" Names
The names for the suppressor shells need to be more unique. For example,
the Stepping action looks for energy deposits in the volume
"right_suppressor_log" to see if the right suppressor fired. The name
for the shell is "shell_for_right_suppressor_log",
"right_suppressor_log" is a subset of that, and it will trigger the find
volume name function. Energy deposits in the Al shell should not fire
the right suppressor.