Added "stepTracker" output
This new functionality outputs step information (event#, step#, det#,
cry#, pos, etc) to a root tuple tree. To activate this code set
stepTrackerBool = true; at the top of EventAction.cc. It is currently
set to false because outputting all this information is computationally
expensive. We don't want to do it unless we need it.